Switch from `prepublish` to `prepare`
As described in the [NPM documentation](https://docs.npmjs.com/misc/scripts#prepublish-and-prepare), the `prepublish` script is deprecated for this use-case. Instead, `prepare` should be used for essential build tasks.

In practice, this makes it possible to depend on `react-native-webview` via git. If I contribute a new feature, I can use the feature right away by depending on its git branch, without waiting for upstream to publish, or even approve my pull request. This is great for experimentation.