Fix `originWhitelist` escape vulnerablity
If the user specifies a specific domain for the origin whitelist, such as `https://example.co`, any domain that starts with this pattern will be valid, including `https://example.com` or `https://example.co.malicious-site.com`.

Fix this by adding a `$` to the end of the regular expression, requiring the domain to match exactly.